Added Unit Test to Handlers

+++ Library.Application/Features/Authors/Command/DeleteAuthor/DeleteAuthorCommandHandler.cs
@@ -1,4 +1,6 @@
-﻿using Library.Application.Interfaces.Persistence;
+﻿using Library.Application.Exceptions;
+using Library.Application.Interfaces.Persistence;
+using Library.Domain.Entities;
 using MediatR;
 
 namespace Library.Application.Features.Authors.Command.DeleteAuthor
@@ -18,6 +20,10 @@ namespace Library.Application.Features.Authors.Command.DeleteAuthor
         public async Task Handle(DeleteAuthorCommand request, CancellationToken cancellationToken)
         {
             var currentAuthor = await _authorRepository.GetAuthorAsync(request.AuthorId, cancellationToken);
+            if (currentAuthor == null)
+            {
+                throw new NotFoundException(nameof(Author), request.AuthorId);
+            }
             _authorRepository.DeleteAuthorSync(currentAuthor);
             await _unitOfWork.SaveChangesAsync(cancellationToken);
         }
